*Hopefully* fixed bug that was preventing user from submitting papers after removing previously submitted papers.
What was happening was that even though a newly submitted paper (through
the GUI) was being removed from the GUI, it wasn't always being removed
from the actual list in the model. I think this may have had to do with
the submitPaper not always submitting the correct version of the author
map (for some strange reason). Seems to work so far, but will be writing
more unit tests